requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
qrcode[pil]==7.4.2
requests==2.32.5
python-dotenv==1.2.1
gTTS==2.5.0
razorpay==1.4.1
